requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
db-dtypes==1.3.0
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
docstring_parser==0.16
executing==2.0.1
frozendict==2.4.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-aiplatform==1.62.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.32
langchain-google-vertexai==1.0.8
langchain-groq==0.1.6
langchain-openai==0.1.19
langchain-text-splitters==0.2.2
langsmith==0.1.93
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.37.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
streamlit==1.37.0
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.41
